chore: apply clippy's mechanical fixes
`cargo clippy --fix` across the workspace, with unused_imports,
unused_variables and dead_code held back -- those three strip items that
are only referenced from #[cfg(test)] modules, which breaks the test build
while the lib build looks fine.

The rest is semantics-preserving: div_ceil in place of (x + n - 1) / n
(which also removes an overflow hazard), let-chains collapsing nested
`if let`, clamp/contains/assign-op idioms, and dropped `mut` on bindings
that are never mutated.

Same 2807 tests and 93 doc tests pass.

+++ crates/dpb-encoders/src/eeg.rs
@@ -932,7 +932,7 @@ fn encode_band_power(
         let power = estimate_band_power(window, sample_rate as f32, config.low_freq, config.high_freq);
 
         let power_value = if config.use_log_power {
-            (power + 1e-10).ln() as f32
+            (power + 1e-10).ln()
         } else {
             power
         };
